requirements.txt
actionlib==1.11.12
angles==1.9.11
attrdict==2.0.0
beautifulsoup4==4.4.1
bleach==1.5.0
bondpy==1.8.1
camera-calibration==1.12.22
camera-calibration-parsers==1.11.13
catkin==0.7.11
catkin-pkg==0.4.1
chardet==2.3.0
click==6.7
cv-bridge==1.12.7
cycler==0.10.0
decorator==4.2.1
diagnostic-analysis==1.9.2
diagnostic-common-diagnostics==1.9.2
diagnostic-updater==1.9.2
docutils==0.14
dynamic-reconfigure==1.5.48
entrypoints==0.2.3
eventlet==0.19.0
Flask==0.11.1
gazebo-plugins==2.5.14
gazebo-ros==2.5.14
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.10
gennodejs==2.0.1
genpy==0.6.7
greenlet==0.4.13
h5py==2.6.0
html5lib==0.9999999
image-geometry==1.12.7
interactive-markers==1.11.3
ipykernel==4.8.2
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.2.0
itsdangerous==0.24
jedi==0.11.1
Jinja2==2.10
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.0.8
kiwisolver==1.0.1
laser-geometry==1.6.4
lxml==3.5.0
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
message-filters==1.12.13
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
notebook==5.4.1
numpy==1.13.1
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
pexpect==4.4.0
pickleshare==0.7.4
Pillow==3.1.2
pluginlib==1.11.2
prompt-toolkit==1.0.15
protobuf==3.5.2.post1
ptyprocess==0.5.2
pycurl==7.43.0
Pygments==2.2.0
pygobject==3.20.0
pyparsing==2.2.0
python-apt==1.1.0b1+ubuntu0.16.4.1
python-box==3.1.1
python-dateutil==2.7.2
python-engineio==2.0.4
python-qt-binding==0.3.3
python-socketio==1.6.1
pytz==2018.3
PyYAML==3.12
pyzmq==17.0.0
qt-dotgraph==0.3.8
qt-gui==0.3.8
qt-gui-cpp==0.3.8
qt-gui-py-common==0.3.8
qtconsole==4.3.1
requests==2.9.1
resource-retriever==1.12.3
rosbag==1.12.13
rosboost-cfg==1.14.3
rosclean==1.14.3
roscreate==1.14.3
rosgraph==1.12.13
roslaunch==1.12.13
roslib==1.14.3
roslint==0.11.0
roslz4==1.12.13
rosmake==1.14.3
rosmaster==1.12.13
rosmsg==1.12.13
rosnode==1.12.13
rosparam==1.12.13
rospkg==1.1.4
rospy==1.12.13
rosservice==1.12.13
rostest==1.12.13
rostopic==1.12.13
rosunit==1.14.3
roswtf==1.12.13
rqt-action==0.4.9
rqt-bag==0.4.11
rqt-bag-plugins==0.4.11
rqt-console==0.4.8
rqt-dep==0.4.8
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.11
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.4.9
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.7
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.5.8
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.15
scipy==0.19.1
Send2Trash==1.5.0
sensor-msgs==1.12.5
simplegeneric==0.8.1
six==1.10.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.1
ssh-import-id==5.5
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
terminado==0.8.1
testpath==0.3.1
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.17
tf2-kdl==0.5.17
tf2-py==0.5.17
tf2-ros==0.5.17
topic-tools==1.12.13
tornado==5.0.1
traitlets==4.3.2
unattended-upgrades==0.1
urllib3==1.13.1
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.2.0
xacro==1.11.2
